+++ aidial_adapter_bedrock/dial_api/embedding_inputs.py
