+++ aidial_adapter_bedrock/embedding/amazon/titan_image.py
@@ -14,6 +14,9 @@ from aidial_sdk.embeddings.request import EmbeddingsRequest
 from pydantic import BaseModel
 
 from aidial_adapter_bedrock.bedrock import Bedrock
+from aidial_adapter_bedrock.dial_api.embedding_inputs import (
+    collect_embedding_inputs,
+)
 from aidial_adapter_bedrock.dial_api.response import make_embeddings_response
 from aidial_adapter_bedrock.dial_api.storage import (
     FileStorage,
@@ -25,7 +28,6 @@ from aidial_adapter_bedrock.embedding.embeddings_adapter import (
     EmbeddingsAdapter,
 )
 from aidial_adapter_bedrock.embedding.encoding import vector_to_base64
-from aidial_adapter_bedrock.embedding.inputs import collect_embedding_inputs
 from aidial_adapter_bedrock.embedding.validation import (
     validate_embeddings_request,
 )
