+++ aidial_adapter_bedrock/embedding/amazon/titan_text.py
@@ -12,15 +12,15 @@ from aidial_sdk.embeddings import Usage
 from aidial_sdk.embeddings.request import EmbeddingsRequest
 
 from aidial_adapter_bedrock.bedrock import Bedrock
+from aidial_adapter_bedrock.dial_api.embedding_inputs import (
+    collect_embedding_inputs_no_attachments,
+)
 from aidial_adapter_bedrock.dial_api.response import make_embeddings_response
 from aidial_adapter_bedrock.embedding.amazon.base import call_embedding_model
 from aidial_adapter_bedrock.embedding.embeddings_adapter import (
     EmbeddingsAdapter,
 )
 from aidial_adapter_bedrock.embedding.encoding import vector_to_base64
-from aidial_adapter_bedrock.embedding.inputs import (
-    collect_embedding_inputs_no_attachments,
-)
 from aidial_adapter_bedrock.embedding.validation import (
     validate_embeddings_request,
 )
